op-chain-ops: cleanup foundry package
Moves the foundry related code into the foundry package.
We have some code for reading the allocs that foundry writes
to disk with `vm.dumpState(string)`. It makes more sense to
have these functions and types in the `foundry` package
rather than the `genesis` package.
